feat: printWinner 메서드 ResultView 클래스로 이동
PlayRacingGame 클래스에 존재한 printWinner 메서드를 ResultView로 이동
