+++ src/main/java/racinggame/view/ResultView.java
@@ -1,7 +1,9 @@
 package racinggame.view;
 
+import java.util.ArrayList;
 import java.util.List;
 import racinggame.domain.Car;
+import racinggame.domain.Winner;
 
 public class ResultView {
 
@@ -15,5 +17,17 @@ public class ResultView {
             }
             System.out.println();
         }
+
+
+    public void printWinner(List<Winner> winners) {
+        List<String> winnerNameList = new ArrayList<>();
+        String winnerNames = "";
+        System.out.print("최종 우승자: ");
+        for (Winner winner : winners) {
+            winnerNameList.add(winner.getWinnerName());
+        }
+        winnerNames = String.join(", ", winnerNameList);
+        System.out.println(winnerNames);
+    }
 
 }
